requirements.txt
flask~=3.0.2
requests~=2.31.0
beautifulsoup4~=4.12.3
openai~=1.16.2
pinecone-client
flask-cors
python-dotenv~=1.0.1
pip~=21.2.4
distro~=1.9.0
wheel~=0.37.0
Jinja2~=3.1.3
future~=0.18.2
certifi~=2024.2.2
urllib3~=2.2.1
tqdm~=4.66.2
idna~=3.6
setuptools~=58.0.4
zipp~=3.18.1